refactor(dataentry): extract views cluster to viewsHandler; delete dead ungated nav path (TKT-R68TV8)
Fixes #1043
