+++ internal/dataentry/api_v1.go
@@ -96,8 +96,8 @@ func (a *App) registerAPIV1Routes(mux *http.ServeMux) {
 	mux.HandleFunc("/api/v1/_theme/logo", a.handleAPIThemeLogo)
 	mux.HandleFunc("/api/v1/_theme/export", a.handleAPIThemeExport)
 	mux.HandleFunc("/api/v1/_theme/import", a.handleAPIThemeImport)
-	mux.HandleFunc("/api/v1/_sidepanel/", a.handleV1SidePanel)
-	mux.HandleFunc("/api/v1/_sidebar", a.handleV1Sidebar)
+	mux.HandleFunc("/api/v1/_sidepanel/", a.views.handleV1SidePanel)
+	mux.HandleFunc("/api/v1/_sidebar", a.views.handleV1Sidebar)
 	mux.HandleFunc("/api/v1/_conflicts", a.handleV1Conflicts)
 	mux.HandleFunc("/api/v1/_conflicts/", a.handleV1ConflictRoutes)
 	mux.HandleFunc("/api/v1/_documents/", a.handleV1Documents)
@@ -108,7 +108,7 @@ func (a *App) registerAPIV1Routes(mux *http.ServeMux) {
 	mux.HandleFunc("/api/v1/_commands", a.handleV1Commands)
 	mux.HandleFunc("/api/v1/_transforms", a.export.handleV1Transforms)
 	mux.HandleFunc("/api/v1/_templates/", a.handleV1Templates)
-	mux.HandleFunc("/api/v1/_views/", a.handleV1Views)
+	mux.HandleFunc("/api/v1/_views/", a.views.handleV1Views)
 	mux.HandleFunc("/api/v1/_action/", a.write.handleV1Action)
 	mux.HandleFunc("/api/v1/_apps/", a.handleV1App)
 
@@ -1214,7 +1214,7 @@ func (a *App) handleV1SchemaRoutes(w http.ResponseWriter, r *http.Request) {
 // resolveRelationWidgets returns a copy of rels with any empty Widget set to
 // "cards" when the relation type has edge properties/content. Shared by the
 // flat and wizard-step relation lists in handleV1Config.
-func (a *App) resolveRelationWidgets(s *Schema, rels []dataentryconfig.FormRelation) []dataentryconfig.FormRelation {
+func resolveRelationWidgets(s *Schema, rels []dataentryconfig.FormRelation) []dataentryconfig.FormRelation {
 	resolved := make([]dataentryconfig.FormRelation, len(rels))
 	copy(resolved, rels)
 	for i := range resolved {
@@ -1258,12 +1258,12 @@ func (a *App) handleV1Config(w http.ResponseWriter, r *http.Request) {
 	forms := make(map[string]dataentryconfig.Form, len(s.Cfg.Forms))
 	for id, form := range s.Cfg.Forms {
 		f := form
-		f.Relations = a.resolveRelationWidgets(s, f.Relations)
+		f.Relations = resolveRelationWidgets(s, f.Relations)
 		if len(f.Steps) > 0 {
 			steps := make([]dataentryconfig.FormStep, len(f.Steps))
 			copy(steps, f.Steps)
 			for i := range steps {
-				steps[i].Relations = a.resolveRelationWidgets(s, steps[i].Relations)
+				steps[i].Relations = resolveRelationWidgets(s, steps[i].Relations)
 			}
 			f.Steps = steps
 		}
@@ -1969,307 +1969,7 @@ func writeV1Error(w http.ResponseWriter, r *http.Request, status int, errType, t
 
 // --- Side Panel API ---
 
-// handleV1SidePanel handles GET /api/v1/_sidepanel/{formId}/{entityId}.
-func (a *App) handleV1SidePanel(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		writeV1Error(w, r, http.StatusMethodNotAllowed, "method_not_allowed", "Method not allowed", "")
-		return
-	}
-
-	// Parse path: /api/v1/_sidepanel/{formId}/{entityId}
-	path := strings.TrimPrefix(r.URL.Path, "/api/v1/_sidepanel/")
-	parts := strings.SplitN(path, "/", 2)
-	if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-		writeV1Error(w, r, http.StatusBadRequest, "invalid_path", "Path must be /_sidepanel/{formId}/{entityId}", "")
-		return
-	}
-
-	formID := parts[0]
-	entityID := parts[1] // Get form config
-	s := a.State()
-	form, ok := s.Cfg.Forms[formID]
-	if !ok {
-		writeV1Error(w, r, http.StatusNotFound, "form_not_found", "Form not found", "")
-		return
-	}
-
-	// Check if form has side panel
-	if form.SidePanel == nil {
-		writeV1JSON(w, http.StatusOK, []v1.SidePanelSection{})
-		return
-	}
-
-	// ACL gate (TKT-6N9O1Y): the side panel reveals the entry entity and its
-	// traversal neighbors. Gate the entry read BEFORE getEntity/executeSidePanel
-	// so a principal who cannot read it gets a 404 indistinguishable from a
-	// missing id, and the traversal never runs for a denied caller.
-	if !a.gateReadOrNotFound(w, r, form.EntityType, entityID) {
-		return
-	}
-
-	// Get the entry entity
-	entry, found := a.reader.getEntity(r.Context(), entityID)
-	if !found {
-		writeV1Error(w, r, http.StatusNotFound, "entity_not_found", "Entity not found", "")
-		return
-	}
-
-	// Execute side panel traversal
-	sections := a.executeSidePanel(r.Context(), form.SidePanel, entityID, form.EntityType)
-	if sections == nil {
-		writeV1JSON(w, http.StatusOK, []v1.SidePanelSection{})
-		return
-	}
-
-	// Build a synthetic ViewConfig to resolve add/link buttons
-	viewConfig := ViewConfig{
-		Entry:    ViewEntry{Type: form.EntityType},
-		Traverse: form.SidePanel.Traverse,
-		Sections: form.SidePanel.Sections,
-	}
-	a.resolveSectionButtonsWithTraverse(viewConfig, sections, entry)
-
-	// Convert to API response format
-	result := make([]v1.SidePanelSection, 0, len(sections))
-	for _, sec := range sections {
-		apiSec := v1.SidePanelSection{
-			Heading:      sec.Heading,
-			SectionID:    sec.SectionID,
-			Display:      sec.Display,
-			IsEmpty:      sec.IsEmpty,
-			EmptyMessage: sec.EmptyMessage,
-		}
-
-		// Convert fields
-		for _, f := range sec.Fields {
-			apiSec.Fields = append(apiSec.Fields, v1.SectionField(f))
-		}
-
-		// Convert entities
-		for _, e := range sec.Entities {
-			apiEnt := v1.SidePanelEntity{
-				ID:         e.ID,
-				Title:      e.Title,
-				Type:       e.Type,
-				EditFormID: e.EditFormID,
-				Content:    e.Content,
-				HasContent: e.HasContent,
-			}
-			for _, f := range e.Fields {
-				apiEnt.Fields = append(apiEnt.Fields, v1.SectionField(f))
-			}
-			apiSec.Entities = append(apiSec.Entities, apiEnt)
-		}
-
-		// Convert add/link info
-		if sec.AddInfo != nil {
-			apiSec.AddInfo = &v1.ViewAddInfo{
-				Relation: sec.AddInfo.Relation,
-				LinkAs:   sec.AddInfo.LinkAs,
-				PeerID:   sec.AddInfo.PeerID,
-			}
-			for _, t := range sec.AddInfo.Targets {
-				apiSec.AddInfo.Targets = append(apiSec.AddInfo.Targets, v1.ViewAddTarget(t))
-			}
-		}
-		if sec.LinkInfo != nil {
-			apiSec.LinkInfo = &v1.ViewLinkInfo{
-				Relation:    sec.LinkInfo.Relation,
-				LinkAs:      sec.LinkInfo.LinkAs,
-				PeerID:      sec.LinkInfo.PeerID,
-				EntityTypes: sec.LinkInfo.EntityTypes,
-			}
-		}
-
-		result = append(result, apiSec)
-	}
-
-	writeV1JSON(w, http.StatusOK, result)
-}
-
 // --- Sidebar API ---
-
-// handleV1Sidebar returns denormalized sidebar data with entity counts.
-func (a *App) handleV1Sidebar(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		writeV1Error(w, r, http.StatusMethodNotAllowed, "method_not_allowed", "Method not allowed", "")
-		return
-	}
-	s := a.State()
-
-	counts := sidebarCounts{
-		filterCache: make(map[string]int),
-		app:         a,
-	}
-
-	// Build navigation with counts
-	navigation := make([]v1.SidebarGroup, 0)
-
-	for _, entry := range s.Cfg.Navigation {
-		if entry.IsGroup() {
-			group := v1.SidebarGroup{
-				Group:     entry.Group,
-				Collapsed: entry.Collapsed,
-				Items:     make([]v1.SidebarItem, 0),
-			}
-			for _, item := range entry.Items {
-				sidebarItem := a.navEntryToSidebarItem(r.Context(), item, counts)
-				group.Items = append(group.Items, sidebarItem)
-			}
-			navigation = append(navigation, group)
-		} else {
-			// Top-level item without group
-			item := a.navEntryToSidebarItem(r.Context(), entry, counts)
-			navigation = append(navigation, v1.SidebarGroup{
-				Items: []v1.SidebarItem{item},
-			})
-		}
-	}
-
-	resp := v1.SidebarResponse{
-		App: v1.AppConfig{
-			Name:        s.Cfg.App.Name,
-			Description: s.Cfg.App.Description,
-		},
-		Navigation: navigation,
-	}
-	resp.LogoURL = a.logo.URL()
-
-	writeV1JSON(w, http.StatusOK, resp)
-}
-
-// sidebarCounts caches sidebar entity counts, applying list- or kanban-
-// level filters when present. Every count flows through the ACL read
-// scope (TKT-VMD8) — one code path regardless of NopACL vs. Declarative
-// (RR-2O27), so the sidebar can never disagree with the list it links
-// to. filterCache is a within-request memo keyed by list/kanban id; it
-// is safe precisely because a sidebarCounts value lives for one request
-// (one principal) — a longer-lived cache would alias counts across
-// principals (RR-BZ4M).
-type sidebarCounts struct {
-	filterCache map[string]int // key: "list:<id>" or "kanban:<id>"
-	app         *App
-}
-
-// listCount returns the entity count for the given list, applying any
-// configured filters. Results are cached per call.
-func (c *sidebarCounts) listCount(ctx context.Context, listID string, list dataentryconfig.List) int {
-	key := "list:" + listID
-	if n, ok := c.filterCache[key]; ok {
-		return n
-	}
-	n := c.countWithFilters(ctx, list.EntityType, list.Filters)
-	c.filterCache[key] = n
-	return n
-}
-
-// kanbanCount returns the entity count for the given kanban, applying
-// any configured filters. Results are cached per call.
-func (c *sidebarCounts) kanbanCount(ctx context.Context, kanbanID string, kanban dataentryconfig.Kanban) int {
-	key := "kanban:" + kanbanID
-	if n, ok := c.filterCache[key]; ok {
-		return n
-	}
-	n := c.countWithFilters(ctx, kanban.EntityType, kanban.Filters)
-	c.filterCache[key] = n
-	return n
-}
-
-// countWithFilters returns the count of entities of the given type that
-// are visible to the requesting principal AND pass the supplied config
-// filters. Ordering is ACL → config filter → count (TKT-VMD8 AC7).
-//
-// Without config filters the count comes straight from GraphCount —
-// identical cost to the old Store.CountEntities for the AllowAll case.
-// With config filters the visible entities are loaded and filtered
-// in-memory; performance scales with the visible-set size (RR-REQW —
-// for visible sets >10k prefer pre-filtering via entity_type in nav
-// config, or file the follow-up that pushes filters into GraphQuery).
-//
-// Backend errors degrade to 0 with a warning — parity with the old
-// CountEntities error path: a broken sidebar count must not take the
-// whole sidebar down, and the list endpoint surfaces the real error.
-//
-// ReadQuery (one member-of walk reuse via the request-scoped
-// acl.Request) and the GraphQuery/GraphCount run once per nav item —
-// two lists over the same type recompute rather than share. Accepted:
-// filterCache keys on list/kanban id, not (type, filters); a
-// (type, filters)-keyed memo is the obvious upgrade if sidebar
-// latency ever warrants it.
-func (c *sidebarCounts) countWithFilters(
-	ctx context.Context, entityType string, filters []dataentryconfig.FilterConfig,
-) int {
-	rqr := readGateFromContext(ctx).ReadQuery(ctx, entityType)
-	if rqr.DenyAll {
-		return 0
-	}
-	q := store.GraphQuery{EntityType: entityType}
-	if rqr.Query != nil {
-		q = *rqr.Query
-	}
-
-	if len(filters) == 0 {
-		matched, _, err := c.app.Services().Store.GraphCount(ctx, q)
-		if err != nil {
-			slog.Warn("sidebar: GraphCount failed; count degraded to 0",
-				"entity_type", entityType, "error", err)
-			return 0
-		}
-		return matched
-	}
-
-	var entities []*entityPkg.Entity
-	for e, err := range c.app.Services().Store.GraphQuery(ctx, q) {
-		if err != nil {
-			slog.Warn("sidebar: GraphQuery failed; count degraded to 0",
-				"entity_type", entityType, "error", err)
-			return 0
-		}
-		entities = append(entities, e)
-	}
-	return len(applyFilters(entities, filters))
-}
-
-// navEntryToSidebarItem converts a navigation entry to a sidebar item with count.
-func (a *App) navEntryToSidebarItem(
-	ctx context.Context, entry dataentryconfig.NavigationEntry, counts sidebarCounts,
-) v1.SidebarItem {
-	s := a.State()
-	item := v1.SidebarItem{
-		Label: entry.Label,
-	}
-
-	switch {
-	case entry.List != "":
-		item.Href = "/list/" + entry.List
-		item.Icon = "list"
-		if list, ok := s.Cfg.Lists[entry.List]; ok {
-			count := counts.listCount(ctx, entry.List, list)
-			item.Count = &count
-		}
-	case entry.Kanban != "":
-		item.Href = "/kanban/" + entry.Kanban
-		item.Icon = "kanban"
-		if kanban, ok := s.Cfg.Kanbans[entry.Kanban]; ok {
-			count := counts.kanbanCount(ctx, entry.Kanban, kanban)
-			item.Count = &count
-		}
-	case entry.Dashboard:
-		item.Href = "/"
-		item.Icon = "dashboard"
-	case entry.Search:
-		item.Href = "/search"
-		item.Icon = "search"
-	case entry.Settings:
-		item.Href = "/settings"
-		item.Icon = "settings"
-	case entry.Action != "":
-		item.Action = entry.Action
-		// Href stays empty — frontend renders this as a button
-	}
-
-	return item
-}
 
 // --- Conflicts API ---
 
@@ -2637,185 +2337,4 @@ func sectionEntityToV1(e SectionEntityData) v1.ViewEntity {
 		v1Ent.FieldAffordances = &fa
 	}
 	return v1Ent
-}
-
-// handleV1Views handles GET /api/v1/_views/{entityType}/{entityId}.
-// Returns JSON with executed view data including entry and sections.
-//
-// View configs are looked up by entry.type. When no explicit ViewConfig
-// is registered for entityType, a default is synthesized from the
-// metamodel (see buildDefaultViewConfig) and executed through the same
-// pipeline so the response shape is identical.
-//
-//nolint:gocognit,funlen // routes the views sub-API over method and path shape; each branch is a distinct view endpoint, not shared logic to factor out.
-func (a *App) handleV1Views(w http.ResponseWriter, r *http.Request) {
-	if r.Method != http.MethodGet {
-		writeV1Error(w, r, http.StatusMethodNotAllowed, "method_not_allowed", "Method not allowed", "")
-		return
-	}
-
-	// Parse path: /api/v1/_views/{entityType}/{entityId}
-	path := strings.TrimPrefix(r.URL.Path, "/api/v1/_views/")
-	parts := strings.SplitN(path, "/", 2)
-	if len(parts) != 2 || parts[0] == "" || parts[1] == "" {
-		writeV1Error(w, r, http.StatusBadRequest, "invalid_path", "Path must be /_views/{entityType}/{entityId}", "")
-		return
-	}
-
-	entityType, entityID := parts[0], parts[1]
-	s := a.State()
-
-	if _, ok := s.Meta.GetEntityDef(entityType); !ok {
-		writeV1Error(w, r, http.StatusNotFound, "entity_type_not_found", "Entity type not found", entityType)
-		return
-	}
-
-	// ACL gate (TKT-BNX2PN): _views is an entity-read chokepoint just like
-	// GET /{plural}/{id} — it serves _title, properties, and content body via
-	// executeView + serializeEntityForWire. Gate BEFORE executeView so a hidden
-	// id is indistinguishable from a missing one (404, no oracle) and the view
-	// pipeline never runs for a denied principal.
-	if !a.gateReadOrNotFound(w, r, entityType, entityID) {
-		return
-	}
-
-	viewCfg, ok := findViewByEntityType(s.Cfg.Views, entityType)
-	if !ok {
-		viewCfg, ok = buildDefaultViewConfig(s.Meta, entityType)
-		if !ok {
-			// Cannot happen — entity type already validated above —
-			// but handled defensively to keep the contract clear.
-			writeV1Error(w, r, http.StatusNotFound, "entity_type_not_found", "Entity type not found", entityType)
-			return
-		}
-	}
-
-	// Execute view
-	result, err := a.executeView(r.Context(), viewCfg, entityID)
-	if err != nil {
-		writeV1Error(w, r, http.StatusUnprocessableEntity, "view_execution_failed", "View execution failed", err.Error())
-		return
-	}
-
-	// Build sections
-	sections := a.buildSections(r.Context(), viewCfg.Sections, result)
-
-	// Build response
-	entityDef := s.Meta.Entities[result.Entry.Type]
-	plural := entityDef.GetPlural(result.Entry.Type)
-
-	entryRels := a.reader.outgoingRelations(r.Context(), result.Entry.ID)
-	resp := v1.ViewResponse{
-		Entry:    a.serializer.forWire(r.Context(), result.Entry, entryRels, a.Meta(), plural),
-		Sections: make([]v1.ViewSection, 0, len(sections)),
-	}
-
-	for _, sec := range sections {
-		v1Sec := v1.ViewSection{
-			Heading:      sec.Heading,
-			SectionID:    sec.SectionID,
-			Display:      sec.Display,
-			IsEmpty:      sec.IsEmpty,
-			EmptyMessage: sec.EmptyMessage,
-			IsGrouped:    sec.IsGrouped,
-			Content:      sec.Content,
-			HasContent:   sec.HasContent,
-		}
-
-		// Convert fields
-		for _, f := range sec.Fields {
-			v1Sec.Fields = append(v1Sec.Fields, v1.SectionField(f))
-		}
-
-		// Convert entities
-		for _, e := range sec.Entities {
-			v1Sec.Entities = append(v1Sec.Entities, sectionEntityToV1(e))
-		}
-
-		// Convert columns
-		for _, col := range sec.Columns {
-			v1Sec.Columns = append(v1Sec.Columns, v1.ViewColumn{
-				Property: col.Property,
-				Label:    col.Label,
-				Relation: col.Relation,
-				Link:     col.Link,
-			})
-		}
-
-		// Convert rows
-		for _, row := range sec.Rows {
-			v1Row := v1.ViewRow{
-				EntityID:   row.EntityID,
-				EntityType: row.EntityType,
-				EditFormID: row.EditFormID,
-				Content:    row.Content,
-			}
-			for _, cell := range row.Cells {
-				v1Row.Cells = append(v1Row.Cells, v1.ViewCell(cell))
-			}
-			v1Sec.Rows = append(v1Sec.Rows, v1Row)
-		}
-
-		// Convert groups
-		for _, grp := range sec.Groups {
-			v1Grp := v1.ViewGroup{
-				GroupName: grp.GroupName,
-			}
-			for _, row := range grp.Rows {
-				v1Row := v1.ViewRow{
-					EntityID:   row.EntityID,
-					EntityType: row.EntityType,
-					EditFormID: row.EditFormID,
-					Content:    row.Content,
-				}
-				for _, cell := range row.Cells {
-					v1Row.Cells = append(v1Row.Cells, v1.ViewCell(cell))
-				}
-				v1Grp.Rows = append(v1Grp.Rows, v1Row)
-			}
-			for _, e := range grp.Entities {
-				v1Grp.Entities = append(v1Grp.Entities, sectionEntityToV1(e))
-			}
-			v1Sec.Groups = append(v1Sec.Groups, v1Grp)
-		}
-
-		resp.Sections = append(resp.Sections, v1Sec)
-	}
-
-	resp.Mentions = collectMentions(
-		r.Context(), a.store, a.viewReader, s.Meta, viewContentBlobs(result.Entry, sections)...)
-
-	writeV1JSON(w, http.StatusOK, resp)
-}
-
-// viewContentBlobs gathers every markdown body that will be rendered by
-// the SPA for a single view response: the entry's content, every section's
-// own content, and every entity card's content (sections with display
-// "content"/"cards" surface related entities, each carrying its own
-// `Content` markdown that EntityDetail.vue renders with the same
-// `refResolver`). Used to scope the mentions scan to text the user
-// actually sees on this screen.
-func viewContentBlobs(entry *entityPkg.Entity, sections []SectionData) []string {
-	blobs := make([]string, 0, 1+len(sections))
-	if entry != nil && entry.Content != "" {
-		blobs = append(blobs, entry.Content)
-	}
-	for _, sec := range sections {
-		if sec.HasContent && sec.Content != "" {
-			blobs = append(blobs, sec.Content)
-		}
-		for _, ent := range sec.Entities {
-			if ent.HasContent && ent.Content != "" {
-				blobs = append(blobs, ent.Content)
-			}
-		}
-		for _, grp := range sec.Groups {
-			for _, ent := range grp.Entities {
-				if ent.HasContent && ent.Content != "" {
-					blobs = append(blobs, ent.Content)
-				}
-			}
-		}
-	}
-	return blobs
 }
