+++ internal/dataentry/views.go
@@ -17,8 +17,8 @@ type viewResult struct {
 }
 
 // executeView runs a view's traversal rules and returns the result.
-func (a *App) executeView(ctx context.Context, view ViewConfig, entryID string) (*viewResult, error) {
-	entry, err := a.store.GetEntity(ctx, entryID)
+func (h *viewsHandler) executeView(ctx context.Context, view ViewConfig, entryID string) (*viewResult, error) {
+	entry, err := h.store.GetEntity(ctx, entryID)
 	if err != nil {
 		return nil, fmt.Errorf("entry entity not found: %s", entryID)
 	}
@@ -36,7 +36,7 @@ func (a *App) executeView(ctx context.Context, view ViewConfig, entryID string) 
 	for range maxPasses {
 		before := countViewEntities(result.Collections)
 		for _, rule := range view.Traverse {
-			a.applyViewTraverse(ctx, rule, result)
+			h.applyViewTraverse(ctx, rule, result)
 		}
 		if countViewEntities(result.Collections) == before {
 			break
@@ -62,15 +62,15 @@ func (a *App) executeView(ctx context.Context, view ViewConfig, entryID string) 
 	// presence/absence in a collection still reflects whether it matched a
 	// predicate over that hidden field. The value is redacted; membership is a
 	// one-bit inference channel, not a value disclosure.
-	result.Entry = visibility.Redact(ctx, appRedactor(a), result.Entry)
+	result.Entry = visibility.Redact(ctx, h.redactor(), result.Entry)
 	for name, entities := range result.Collections {
-		result.Collections[name] = a.viewReader.Filter(ctx, entities)
+		result.Collections[name] = h.viewReader.Filter(ctx, entities)
 	}
 
 	return result, nil
 }
 
-func (a *App) applyViewTraverse(ctx context.Context, rule ViewTraverse, result *viewResult) {
+func (h *viewsHandler) applyViewTraverse(ctx context.Context, rule ViewTraverse, result *viewResult) {
 	// Gather source entities
 	var sources []*entity.Entity
 	if rule.From == "*" {
@@ -96,15 +96,15 @@ func (a *App) applyViewTraverse(ctx context.Context, rule ViewTraverse, result *
 			if maxD <= 0 {
 				maxD = maxRecursionDepth
 			}
-			found = append(found, a.traverseViewRecursive(ctx, src.ID, rule, 0, maxD, map[string]bool{})...)
+			found = append(found, h.traverseViewRecursive(ctx, src.ID, rule, 0, maxD, map[string]bool{})...)
 		} else {
-			found = append(found, a.traverseViewOnce(ctx, src.ID, rule)...)
+			found = append(found, h.traverseViewOnce(ctx, src.ID, rule)...)
 		}
 	}
 
 	// Apply where filter if specified
 	if rule.Where != "" {
-		filtered, err := a.filterEntities(found, rule.Where)
+		filtered, err := h.filterEntities(found, rule.Where)
 		if err == nil {
 			found = filtered
 		}
@@ -127,8 +127,8 @@ func (a *App) applyViewTraverse(ctx context.Context, rule ViewTraverse, result *
 	}
 }
 
-func (a *App) traverseViewOnce(ctx context.Context, sourceID string, rule ViewTraverse) []*entity.Entity {
-	st := a.store
+func (h *viewsHandler) traverseViewOnce(ctx context.Context, sourceID string, rule ViewTraverse) []*entity.Entity {
+	st := h.store
 	var out []*entity.Entity
 
 	var relType string
@@ -163,18 +163,18 @@ func (a *App) traverseViewOnce(ctx context.Context, sourceID string, rule ViewTr
 	return out
 }
 
-func (a *App) traverseViewRecursive(
+func (h *viewsHandler) traverseViewRecursive(
 	ctx context.Context, sourceID string, rule ViewTraverse, depth, maxDepth int, visited map[string]bool,
 ) []*entity.Entity {
 	if depth >= maxDepth || visited[sourceID] {
 		return nil
 	}
 	visited[sourceID] = true
-	immediate := a.traverseViewOnce(ctx, sourceID, rule)
+	immediate := h.traverseViewOnce(ctx, sourceID, rule)
 	var all []*entity.Entity
 	all = append(all, immediate...)
 	for _, e := range immediate {
-		all = append(all, a.traverseViewRecursive(ctx, e.ID, rule, depth+1, maxDepth, visited)...)
+		all = append(all, h.traverseViewRecursive(ctx, e.ID, rule, depth+1, maxDepth, visited)...)
 	}
 	return all
 }
@@ -191,13 +191,13 @@ func countViewEntities(collections map[string][]*entity.Entity) int {
 
 // filterEntities filters entities based on a where expression.
 // Supports the "type" pseudo-property to filter by entity type.
-func (a *App) filterEntities(entities []*entity.Entity, whereExpr string) ([]*entity.Entity, error) {
+func (h *viewsHandler) filterEntities(entities []*entity.Entity, whereExpr string) ([]*entity.Entity, error) {
 	f, err := filter.Parse(whereExpr)
 	if err != nil {
 		return nil, fmt.Errorf("invalid where expression: %w", err)
 	}
 
-	s := a.State()
+	s := h.schema()
 	var result []*entity.Entity
 	for _, e := range entities {
 		// Special handling for "type" pseudo-property
